[FLINK-35158][Core/API] Move functional interfaces about consumers from core module to core-api
