+++ flink-core-api/src/main/java/org/apache/flink/util/function/LongFunctionWithException.java
