+++ flink-core-api/src/main/java/org/apache/flink/util/function/ThrowingExceptionUtils.java
@@ -1,0 +1,42 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.flink.util.function;
+
+/** Exception utils only for this package. */
+class ThrowingExceptionUtils {
+
+    /**
+     * Throws the given {@code Throwable} in scenarios where the signatures do not allow you to
+     * throw an arbitrary Throwable. Errors and RuntimeExceptions are thrown directly, other
+     * exceptions are packed into runtime exceptions
+     *
+     * @param t The throwable to be thrown.
+     */
+    static void rethrow(Throwable t) {
+        if (t instanceof Error) {
+            throw (Error) t;
+        } else if (t instanceof RuntimeException) {
+            throw (RuntimeException) t;
+        } else {
+            throw new RuntimeException(t);
+        }
+    }
+
+    private ThrowingExceptionUtils() {}
+}
