+++ flink-core-api/src/main/java/org/apache/flink/util/function/TriFunction.java
