+++ pom.xml
@@ -2365,6 +2365,12 @@ under the License.
 								<!-- MARKER: start exclusions; these will be wiped by tools/releasing/update_japicmp_configuration.sh -->
 								<exclude>org.apache.flink.api.common.functions.Function</exclude>
 								<exclude>org.apache.flink.api.java.functions.KeySelector</exclude>
+								<exclude>org.apache.flink.util.function.FunctionWithException</exclude>
+								<exclude>org.apache.flink.util.function.LongFunctionWithException</exclude>
+								<exclude>org.apache.flink.util.function.RunnableWithException</exclude>
+								<exclude>org.apache.flink.util.function.SerializableFunction</exclude>
+								<exclude>org.apache.flink.util.function.SupplierWithException</exclude>
+								<exclude>org.apache.flink.util.function.ThrowingConsumer</exclude>
 								<!-- MARKER: end exclusions -->
 							</excludes>
 							<accessModifier>public</accessModifier>
